# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Enicilion Platform API
  description: Complete OpenAPI specification for Enicilion Enterprise Java Spring Boot Backend.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local Development Server
paths:
...
  /api/tickets:
    get:
      summary: Retrieve authenticated user's tickets
      tags:
        - Tickets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User tickets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        tierName:
                          type: string
                        status:
                          type: string
                          enum: [booked, paid, checked_in, cancelled]
...
  /api/support/tickets:
    post:
      summary: Submit contact support queries
      tags:
        - Support Helpdesk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category
                - message
              properties:
                category:
                  type: string
                  example: billing
                message:
                  type: string
                  example: Help with ticket payment
                name:
                  type: string
                  example: Jane Doe
                phone:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Support query recorded
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...